Make sure we handle cases where an editable scalar has a value that is an unexpected tag.
This scenario can happen if someone edits the value from, say `__MISSING__` to `__OPTIONAL__` or if we change a field within Hyde from optional to required, for example. Before, we would have have accepted such a value. Now, it will be a validation error, and will corrected to the expected tag during the update operation.